Cleans up/standardizes the ISelectable interface and applies it to Village.
Has the individual ISelectable entities (HexCell, Unit, Village, etc) setup the in-game UI as needed via the OnSelected, OnBlur, and OnUpdateWhileSelected events.  Normalizes the behavoir across all these entities and has Village use the ISelectable interface as well.

Previously the ISelectable interface was too stringent and geared towards entities using the SelectionPanel only, and was not robust to UI changes (the UI fields had to correspond one-to-one to the interface fields for values to populate -- this is replaced with OnUpdateWhileSelected).

Fixes #31.