Adds ability to create tile improvement on a hex tile
Fixes #10 by adding the ability to create a tile imrovement on any hex tile.

Reuses/repurposes the "Create Village" button to also function as a "Build Tile Improvement" button, when a hex cell is selected (vs a player unit) and there is at least one village selected. In future there will be additional restrictions on when this button is shown.

Tile improvement creation functions much like village creation, but its gameobject is slightly different and it does not require sacrifing a unit to create.  Currently there is no ability to select which tile improvememnt is constructed but in future the UI will be expanded to have different options for the various types of buildings.